BG Scrolling, edited some scripts
Created scripts allowing background textures to scroll and paralax, as
well as follow the player. Also made some changes to the camera script
which is now FollowPlayer.cs allowing better control on it's position in
the game and general use between a camera and a scrolling background.

The added materials and textures are purely for testing and debugging
the BG scrolling.